Add read more to phase description
