+++ front/app/containers/ProjectsShowPage/shared/header/ReadMoreWrapper.tsx
@@ -1,0 +1,196 @@
+import React, { memo, useCallback, useState, useEffect } from 'react';
+import { isEmpty } from 'lodash-es';
+
+// components
+import Button from 'components/UI/Button';
+import ReactResizeDetector from 'react-resize-detector';
+
+// hooks
+import { useWindowSize } from '@citizenlab/cl2-component-library';
+
+// i18n
+import T from 'components/T';
+import { FormattedMessage } from 'utils/cl-intl';
+import messages from 'containers/ProjectsShowPage/messages';
+
+// style
+import styled, { useTheme } from 'styled-components';
+import {
+  fontSizes,
+  colors,
+  media,
+  viewportWidths,
+  isRtl,
+} from 'utils/styleUtils';
+import QuillEditedContent from 'components/UI/QuillEditedContent';
+
+import { Multiloc } from 'typings';
+
+const desktopCollapsedContentMaxHeight = 380;
+const mobileCollapsedContentMaxHeight = 180;
+
+const Container = styled.div`
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+
+  ${isRtl`
+    flex-direction: row-reverse;
+  `}
+
+  ${media.smallerThanMaxTablet`
+    flex-direction: column;
+    align-items: stretch;
+    justify-content: flex-start;
+  `}
+`;
+
+const Content = styled.div<{ maxHeight: number }>`
+  position: relative;
+  max-height: ${(props) => props.maxHeight}px;
+  overflow: hidden;
+
+  &.expanded {
+    max-height: none;
+    overflow: visible;
+  }
+`;
+
+const ReadMoreOuterWrapper = styled.div`
+  height: 130px;
+  content: '';
+  display: flex;
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  background: rgb(255, 255, 255);
+  background: linear-gradient(
+    0deg,
+    rgba(255, 255, 255, 1) 30%,
+    rgba(255, 255, 255, 0) 100%
+  );
+`;
+
+const ReadMoreInnerWrapper = styled.div`
+  position: relative;
+  flex: 1;
+`;
+
+const ReadMoreButton = styled(Button)`
+  position: absolute;
+  bottom: 0;
+  left: 0;
+`;
+
+const CollapseButtonWrapper = styled.div`
+  display: flex;
+  justify-content: flex-start;
+  margin-top: 20px;
+`;
+
+const CollapseButton = styled(Button)``;
+
+interface Props {
+  className?: string;
+  fontSize: 'base' | 's' | 'm' | 'l';
+  value: Multiloc;
+}
+
+const ReadMoreWrapper = memo<Props>(({ className, value, fontSize }) => {
+  const theme: any = useTheme();
+  const { windowWidth } = useWindowSize();
+
+  const [expanded, setExpanded] = useState(false);
+  const [contentHeight, setContentHeight] = useState<number | null>(null);
+
+  const smallerThanLargeTablet = windowWidth <= viewportWidths.largeTablet;
+
+  const collapsedContentMaxHeight = smallerThanLargeTablet
+    ? mobileCollapsedContentMaxHeight
+    : desktopCollapsedContentMaxHeight;
+
+  useEffect(() => {
+    setExpanded(false);
+  }, [value]);
+
+  const toggleExpandCollapse = useCallback((event: React.MouseEvent) => {
+    event.preventDefault();
+    setExpanded((expanded) => !expanded);
+  }, []);
+
+  const onResize = (_width: number | undefined, height: number | undefined) => {
+    if (height) {
+      setContentHeight(height);
+    }
+  };
+
+  return (
+    <Container className={`${className || ''} e2e-project-info`}>
+      <Content
+        className={expanded ? 'expanded' : ''}
+        maxHeight={collapsedContentMaxHeight}
+      >
+        {!isEmpty(value) && (
+          <>
+            <ReactResizeDetector handleWidth handleHeight onResize={onResize}>
+              <div id="e2e-project-description">
+                <QuillEditedContent
+                  fontSize={fontSize}
+                  textColor={theme.colorText}
+                  disableTabbing={!expanded}
+                >
+                  <T value={value} supportHtml={true} />
+                </QuillEditedContent>
+              </div>
+            </ReactResizeDetector>
+            {contentHeight &&
+              contentHeight > collapsedContentMaxHeight &&
+              !expanded && (
+                <ReadMoreOuterWrapper>
+                  <ReadMoreInnerWrapper>
+                    <ReadMoreButton
+                      id="e2e-project-description-read-more-button"
+                      buttonStyle="text"
+                      onClick={toggleExpandCollapse}
+                      textDecoration="underline"
+                      textDecorationHover="underline"
+                      textColor={colors.label}
+                      textHoverColor={theme.colorText}
+                      fontWeight="500"
+                      fontSize={`${fontSizes.m}px`}
+                      padding="0"
+                    >
+                      <FormattedMessage {...messages.readMore} />
+                    </ReadMoreButton>
+                  </ReadMoreInnerWrapper>
+                </ReadMoreOuterWrapper>
+              )}
+            {contentHeight &&
+              contentHeight > collapsedContentMaxHeight &&
+              expanded && (
+                <CollapseButtonWrapper>
+                  <CollapseButton
+                    id="e2e-project-description-see-less-button"
+                    buttonStyle="text"
+                    onClick={toggleExpandCollapse}
+                    textDecoration="underline"
+                    textDecorationHover="underline"
+                    textColor={colors.label}
+                    textHoverColor={theme.colorText}
+                    fontWeight="500"
+                    fontSize={`${fontSizes.m}px`}
+                    padding="0"
+                  >
+                    <FormattedMessage {...messages.seeLess} />
+                  </CollapseButton>
+                </CollapseButtonWrapper>
+              )}
+          </>
+        )}
+      </Content>
+    </Container>
+  );
+});
+
+export default ReadMoreWrapper;
