+++ front/app/containers/ProjectsShowPage/timeline/PhaseDescription.tsx
@@ -5,6 +5,7 @@ import { isEmpty } from 'lodash-es';
 // components
 import FileAttachments from 'components/UI/FileAttachments';
 import PhaseTitle from './PhaseTitle';
+import ReadMoreWrapper from 'containers/ProjectsShowPage/shared/header/ReadMoreWrapper';
 
 // hooks
 import useLocalize from 'hooks/useLocalize';
@@ -12,10 +13,8 @@ import usePhase from 'hooks/usePhase';
 import useResourceFiles from 'hooks/useResourceFiles';
 
 // style
-import styled, { useTheme } from 'styled-components';
+import styled from 'styled-components';
 import { defaultCardStyle, media } from 'utils/styleUtils';
-import T from 'components/T';
-import QuillEditedContent from 'components/UI/QuillEditedContent';
 
 const Container = styled.div`
   padding: 30px;
@@ -45,7 +44,6 @@ const PhaseDescription = ({
   className,
   hidden,
 }: Props) => {
-  const theme: any = useTheme();
   const localize = useLocalize();
   const phase = usePhase(phaseId);
   const phaseFiles = useResourceFiles({
@@ -76,12 +74,10 @@ const PhaseDescription = ({
       />
       {!isNilOrError(phase) && descriptionHasContent && (
         <>
-          <QuillEditedContent fontSize="base" textColor={theme.colorText}>
-            <T
-              value={phase?.attributes?.description_multiloc}
-              supportHtml={true}
-            />
-          </QuillEditedContent>
+          <ReadMoreWrapper
+            fontSize="base"
+            value={phase?.attributes?.description_multiloc}
+          />
 
           {!isNilOrError(phaseFiles) && !isEmpty(phaseFiles) && (
             <StyledFileAttachments files={phaseFiles} />
